@@ -12,16 +12,10 @@ buildscript {
         }
     }
     dependencies {
-        classpath group: 'net.minecraftforge.gradle', name: 'ForgeGradle', version: '3.+', changing: true
-        classpath group: 'org.spongepowered', name: 'mixingradle', version: '0.7-SNAPSHOT'
+        classpath group: 'net.minecraftforge.gradle', name: 'ForgeGradle', version: '5.+', changing: true
     }
 }
 apply plugin: 'net.minecraftforge.gradle'
-apply plugin: 'org.spongepowered.mixin'
-
-mixin {
-    add sourceSets.main, "framedcompactdrawers.refmap.json"
-}
 
 sourceSets.main.resources {
     srcDir 'src/generated/resources'
@@ -33,12 +27,12 @@ version = "${mc_part}-${major}.${minor}.${patch}"
 group = "eutros.framedcompactdrawers"
 archivesBaseName = "${mod_id}"
 
-sourceCompatibility = targetCompatibility = compileJava.sourceCompatibility = compileJava.targetCompatibility = '1.8'
+java.toolchain.languageVersion = JavaLanguageVersion.of(16)
 
 println('Java: ' + System.getProperty('java.version') + ' JVM: ' + System.getProperty('java.vm.version') + '(' + System.getProperty('java.vendor') + ') Arch: ' + System.getProperty('os.arch'))
 minecraft {
-    mappings channel: 'snapshot', version: '20200723-1.16.1'
-    // accessTransformer = file('src/main/resources/META-INF/accesstransformer.cfg')
+    mappings channel: 'official', version: '1.17.1'
+    accessTransformer = file('src/main/resources/META-INF/accesstransformer.cfg')
     runs {
         client {
             property 'forge.logging.console.level', 'debug'
@@ -84,14 +78,6 @@ repositories {
 
 dependencies {
     minecraft "net.minecraftforge:forge:${mc_version}-${forge_version}"
-
-    annotationProcessor 'org.spongepowered:mixin:0.8'
-    annotationProcessor 'com.google.code.gson:gson:2.8.0'
-    annotationProcessor 'com.google.guava:guava:21.0'
-    annotationProcessor 'org.ow2.asm:asm-tree:7.2'
-    annotationProcessor 'org.ow2.asm:asm-commons:7.2'
-    annotationProcessor 'org.ow2.asm:asm-util:7.2'
-
     implementation fg.deobf("curse.maven:StorageDrawers-223852:${storagedrawers_version}")
 }
 
@@ -102,8 +88,7 @@ jar {
                 "Specification-Vendor"    : "Eutros",
                 "Implementation-Title"    : project.name,
                 "Implementation-Version"  : "${project.version}",
-                "Implementation-Timestamp": new Date().format("yyyy-MM-dd'T'HH:mm:ssZ"),
-                "MixinConnector"          : "eutros.framedcompactdrawers.MixinConnector"
+                "Implementation-Timestamp": new Date().format("yyyy-MM-dd'T'HH:mm:ssZ")
         ])
     }
 }
