@@ -7,17 +7,17 @@ import com.jaquadro.minecraft.storagedrawers.block.tile.TileEntityDrawersStandar
 import com.jaquadro.minecraft.storagedrawers.core.ModBlocks;
 import eutros.framedcompactdrawers.block.tile.IFramingHolder;
 import eutros.framedcompactdrawers.block.tile.TileDrawersStandardCustom;
-import net.minecraft.block.BlockState;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.item.ItemStack;
-import net.minecraft.tileentity.TileEntity;
-import net.minecraft.util.ActionResultType;
-import net.minecraft.util.Hand;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.BlockRayTraceResult;
-import net.minecraft.util.math.RayTraceResult;
-import net.minecraft.world.IBlockReader;
-import net.minecraft.world.World;
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResult;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.BlockGetter;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.entity.BlockEntity;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.phys.BlockHitResult;
+import net.minecraft.world.phys.HitResult;
 
 public class BlockDrawersStandardCustom extends BlockStandardDrawers {
 
@@ -26,13 +26,13 @@ public class BlockDrawersStandardCustom extends BlockStandardDrawers {
     }
 
     @Override
-    public TileEntityDrawersStandard createTileEntity(BlockState state, IBlockReader world) {
-        return TileDrawersStandardCustom.createEntity(getDrawerCount());
+    public TileEntityDrawersStandard newBlockEntity(BlockPos pos, BlockState state) {
+        return TileDrawersStandardCustom.createEntity(getDrawerCount(), pos, state);
     }
 
     @Override
-    public ItemStack getPickBlock(BlockState state, RayTraceResult target, IBlockReader world, BlockPos pos, PlayerEntity player) {
-        TileEntity tile = world.getTileEntity(pos);
+    public ItemStack getPickBlock(BlockState state, HitResult target, BlockGetter world, BlockPos pos, Player player) {
+        BlockEntity tile = world.getBlockEntity(pos);
         ItemStack stack = super.getPickBlock(state, target, world, pos, player);
         if(tile instanceof IFramingHolder) {
             ((IFramingHolder) tile).writeToTag(stack.getOrCreateTag());
@@ -41,12 +41,12 @@ public class BlockDrawersStandardCustom extends BlockStandardDrawers {
     }
 
     @Override
-    public ActionResultType onBlockActivated(BlockState state, World world, BlockPos pos, PlayerEntity player, Hand hand, BlockRayTraceResult hit) {
-        TileEntity tile = world.getTileEntity(pos);
+    public InteractionResult use(BlockState state, Level world, BlockPos pos, Player player, InteractionHand hand, BlockHitResult hit) {
+        BlockEntity tile = world.getBlockEntity(pos);
         if(tile instanceof IFramingHolder && !((IFramingHolder) tile).getSide().isEmpty()) {
-            return super.onBlockActivated(state, world, pos, player, hand, hit);
+            return super.use(state, world, pos, player, hand, hit);
         }
-        return ActionResultType.PASS;
+        return InteractionResult.PASS;
     }
 
     public void setGeometryData() {
@@ -58,27 +58,19 @@ public class BlockDrawersStandardCustom extends BlockStandardDrawers {
 
     private BlockDrawers getGeometrySource() {
         if(isHalfDepth()) {
-            switch(getDrawerCount()) {
-                case 1:
-                    return ModBlocks.OAK_HALF_DRAWERS_1;
-                case 2:
-                    return ModBlocks.OAK_HALF_DRAWERS_2;
-                case 4:
-                    return ModBlocks.OAK_HALF_DRAWERS_4;
-                default:
-                    throw new IllegalArgumentException("Illegal drawer count.");
-            }
+            return switch (getDrawerCount()) {
+                case 1 -> ModBlocks.OAK_HALF_DRAWERS_1;
+                case 2 -> ModBlocks.OAK_HALF_DRAWERS_2;
+                case 4 -> ModBlocks.OAK_HALF_DRAWERS_4;
+                default -> throw new IllegalArgumentException("Illegal drawer count.");
+            };
         } else {
-            switch(getDrawerCount()) {
-                case 1:
-                    return ModBlocks.OAK_FULL_DRAWERS_1;
-                case 2:
-                    return ModBlocks.OAK_FULL_DRAWERS_2;
-                case 4:
-                    return ModBlocks.OAK_FULL_DRAWERS_4;
-                default:
-                    throw new IllegalArgumentException("Illegal drawer count.");
-            }
+            return switch (getDrawerCount()) {
+                case 1 -> ModBlocks.OAK_FULL_DRAWERS_1;
+                case 2 -> ModBlocks.OAK_FULL_DRAWERS_2;
+                case 4 -> ModBlocks.OAK_FULL_DRAWERS_4;
+                default -> throw new IllegalArgumentException("Illegal drawer count.");
+            };
         }
     }
 
