@@ -2,7 +2,9 @@ package eutros.framedcompactdrawers.block.tile;
 
 import com.jaquadro.minecraft.storagedrawers.block.tile.TileEntitySlave;
 import eutros.framedcompactdrawers.block.ModBlocks;
-import net.minecraft.item.ItemStack;
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.block.state.BlockState;
 import net.minecraftforge.client.model.data.IModelData;
 
 import javax.annotation.Nonnull;
@@ -13,8 +15,8 @@ public class TileSlaveCustom extends TileEntitySlave implements IFramingHolder {
     private ItemStack trim = ItemStack.EMPTY;
     private ItemStack front = ItemStack.EMPTY;
 
-    public TileSlaveCustom() {
-        super(ModBlocks.Tile.slaveCustom);
+    public TileSlaveCustom(BlockPos pos, BlockState state) {
+        super(ModBlocks.Tile.slaveCustom, pos, state);
         injectCustomData(this);
     }
 
