@@ -6,60 +6,64 @@ import eutros.framedcompactdrawers.FramedCompactDrawers;
 import eutros.framedcompactdrawers.block.ModBlocks;
 import eutros.framedcompactdrawers.recipe.FramingRecipe;
 import eutros.framedcompactdrawers.recipe.ModTags;
-import net.minecraft.block.Block;
+import net.minecraft.MethodsReturnNonnullByDefault;
 import net.minecraft.data.DataGenerator;
-import net.minecraft.data.IFinishedRecipe;
-import net.minecraft.data.ShapedRecipeBuilder;
-import net.minecraft.item.Item;
-import net.minecraft.item.crafting.IRecipeSerializer;
-import net.minecraft.item.crafting.Ingredient;
+import net.minecraft.data.recipes.FinishedRecipe;
+import net.minecraft.data.recipes.RecipeProvider;
+import net.minecraft.data.recipes.ShapedRecipeBuilder;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.ItemTags;
-import net.minecraft.util.ResourceLocation;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.crafting.Ingredient;
+import net.minecraft.world.item.crafting.RecipeSerializer;
+import net.minecraft.world.level.block.Block;
 import net.minecraftforge.common.Tags;
-import net.minecraftforge.common.data.ForgeRecipeProvider;
 import org.apache.commons.lang3.tuple.Pair;
 import org.apache.commons.lang3.tuple.Triple;
 
 import javax.annotation.Nullable;
+import javax.annotation.ParametersAreNonnullByDefault;
 import java.util.Arrays;
 import java.util.function.Consumer;
 
-public class FCDRecipeProvider extends ForgeRecipeProvider {
+@ParametersAreNonnullByDefault
+@MethodsReturnNonnullByDefault
+public class FCDRecipeProvider extends RecipeProvider {
 
     public FCDRecipeProvider(DataGenerator generatorIn) {
         super(generatorIn);
     }
 
     @Override
-    protected void registerRecipes(Consumer<IFinishedRecipe> consumer) {
+    protected void buildCraftingRecipes(Consumer<FinishedRecipe> consumer) {
         for (Pair<? extends Block, Tags.IOptionalNamedTag<Item>> pair : ImmutableList.of(
                 Pair.of(ModBlocks.framedCompactDrawer, ModTags.Items.COMPACTING),
                 Pair.of(ModBlocks.framedDrawerController, ModTags.Items.CONTROLLER),
                 Pair.of(ModBlocks.framedSlave, ModTags.Items.SLAVE)
         )) {
-            ShapedRecipeBuilder.shapedRecipe(pair.getLeft())
-                    .patternLine("///")
-                    .patternLine("/D/")
-                    .patternLine("///")
-                    .key('/', Tags.Items.RODS_WOODEN)
-                    .key('D', pair.getRight())
-                    .addCriterion("has_" + pair.getRight().getName().getNamespace(),
-                            hasItem(pair.getRight()))
-                    .addCriterion("has_sticks", hasItem(Tags.Items.RODS_WOODEN))
-                    .setGroup("framed_drawers")
-                    .build(consumer);
+            ShapedRecipeBuilder.shaped(pair.getLeft())
+                    .pattern("///")
+                    .pattern("/D/")
+                    .pattern("///")
+                    .define('/', Tags.Items.RODS_WOODEN)
+                    .define('D', pair.getRight())
+                    .unlockedBy("has_" + pair.getRight().getName().getNamespace(),
+                            has(pair.getRight()))
+                    .unlockedBy("has_sticks", has(Tags.Items.RODS_WOODEN))
+                    .group("framed_drawers")
+                    .save(consumer);
         }
 
-        ShapedRecipeBuilder.shapedRecipe(ModBlocks.framedTrim, 4)
-                .patternLine("/X/")
-                .patternLine("X/X")
-                .patternLine("/X/")
-                .key('/', Tags.Items.RODS_WOODEN)
-                .key('X', ItemTags.PLANKS)
-                .addCriterion("has_sticks", hasItem(Tags.Items.RODS_WOODEN))
-                .addCriterion("has_planks", hasItem(ItemTags.PLANKS))
-                .setGroup("framed_drawers")
-                .build(consumer);
+        ShapedRecipeBuilder.shaped(ModBlocks.framedTrim, 4)
+                .pattern("/X/")
+                .pattern("X/X")
+                .pattern("/X/")
+                .define('/', Tags.Items.RODS_WOODEN)
+                .define('X', ItemTags.PLANKS)
+                .unlockedBy("has_sticks", has(Tags.Items.RODS_WOODEN))
+                .unlockedBy("has_planks", has(ItemTags.PLANKS))
+                .group("framed_drawers")
+                .save(consumer);
 
         for (Triple<? extends Block, Integer, String[]> triple : ImmutableList.of(
                 Triple.of(ModBlocks.framedFullOne, 1, new String[]{
@@ -95,29 +99,29 @@ public class FCDRecipeProvider extends ForgeRecipeProvider {
                 })
         )) {
             ShapedRecipeBuilder srb = ShapedRecipeBuilder
-                    .shapedRecipe(triple.getLeft(), triple.getMiddle())
-                    .patternLine(triple.getRight()[0])
-                    .patternLine(triple.getRight()[1])
-                    .patternLine(triple.getRight()[2])
-                    .key('/', Tags.Items.RODS_WOODEN)
-                    .key('C', Tags.Items.CHESTS_WOODEN)
-                    .addCriterion("has_chests_wooden", hasItem(Tags.Items.CHESTS_WOODEN))
-                    .addCriterion("has_sticks", hasItem(Tags.Items.RODS_WOODEN))
-                    .setGroup("framed_drawers");
+                    .shaped(triple.getLeft(), triple.getMiddle())
+                    .pattern(triple.getRight()[0])
+                    .pattern(triple.getRight()[1])
+                    .pattern(triple.getRight()[2])
+                    .define('/', Tags.Items.RODS_WOODEN)
+                    .define('C', Tags.Items.CHESTS_WOODEN)
+                    .unlockedBy("has_chests_wooden", has(Tags.Items.CHESTS_WOODEN))
+                    .unlockedBy("has_sticks", has(Tags.Items.RODS_WOODEN))
+                    .group("framed_drawers");
 
             if (Arrays.stream(triple.getRight()).anyMatch(s -> s.contains("S"))) {
-                srb.key('S', ItemTags.WOODEN_SLABS)
-                        .addCriterion("has_wooden_slabs", hasItem(ItemTags.WOODEN_SLABS));
+                srb.define('S', ItemTags.WOODEN_SLABS)
+                        .unlockedBy("has_wooden_slabs", has(ItemTags.WOODEN_SLABS));
             }
 
-            srb.build(consumer);
+            srb.save(consumer);
         }
 
-        consumer.accept(new FinishedFramingRecipe("frame_three", Ingredient.fromTag(ModTags.Items.FRAME_TRIPLE), true));
-        consumer.accept(new FinishedFramingRecipe("frame_two", Ingredient.fromTag(ModTags.Items.FRAME_DOUBLE), false));
+        consumer.accept(new FinishedFramingRecipe("frame_three", Ingredient.of(ModTags.Items.FRAME_TRIPLE), true));
+        consumer.accept(new FinishedFramingRecipe("frame_two", Ingredient.of(ModTags.Items.FRAME_DOUBLE), false));
     }
 
-    private static class FinishedFramingRecipe implements IFinishedRecipe {
+    private static class FinishedFramingRecipe implements FinishedRecipe {
 
         private final String path;
         private final Ingredient ingredient;
@@ -130,27 +134,30 @@ public class FCDRecipeProvider extends ForgeRecipeProvider {
         }
 
         @Override
-        public void serialize(JsonObject json) {
-            json.add("ingredient", ingredient.serialize());
+        public void serializeRecipeData(JsonObject json) {
+            json.add("ingredient", ingredient.toJson());
             json.addProperty("includeFront", includeFront);
         }
 
         @Override
-        public ResourceLocation getID() {
+        public ResourceLocation getId() {
             return new ResourceLocation(FramedCompactDrawers.MOD_ID, path);
         }
 
         @Override
-        public IRecipeSerializer<?> getSerializer() {
+        public RecipeSerializer<?> getType() {
             return FramingRecipe.SERIALIZER;
         }
 
+        @Override
         @Nullable
-        public JsonObject getAdvancementJson() {
+        public JsonObject serializeAdvancement() {
             return null;
         }
 
-        public ResourceLocation getAdvancementID() {
+        @Override
+        @Nullable
+        public ResourceLocation getAdvancementId() {
             return new ResourceLocation("");
         }
     }
