@@ -1,17 +1,19 @@
 package eutros.framedcompactdrawers.item;
 
-import net.minecraft.block.Block;
-import net.minecraft.block.BlockState;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.item.BlockItem;
-import net.minecraft.item.ItemStack;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.World;
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.BlockItem;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.state.BlockState;
 
 import javax.annotation.Nullable;
+import javax.annotation.ParametersAreNonnullByDefault;
 
 import static eutros.framedcompactdrawers.item.ItemDrawersCustom.setFrame;
 
+@ParametersAreNonnullByDefault
 public class ItemOtherCustom extends BlockItem {
 
     public ItemOtherCustom(Block block, Properties properties) {
@@ -19,8 +21,8 @@ public class ItemOtherCustom extends BlockItem {
     }
 
     @Override
-    protected boolean onBlockPlaced(BlockPos pos, World worldIn, @Nullable PlayerEntity player, ItemStack stack, BlockState state) {
-        super.onBlockPlaced(pos, worldIn, player, stack, state);
+    protected boolean updateCustomBlockEntityTag(BlockPos pos, Level worldIn, @Nullable Player player, ItemStack stack, BlockState state) {
+        super.updateCustomBlockEntityTag(pos, worldIn, player, stack, state);
         return setFrame(pos, worldIn, stack);
     }
 
